chore: rename crates and update imports pre-release
