--- libworker/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-edition = "2018"
-name    = "libworker"
-authors = ["Cloudflare Workers Team <workers@cloudflare.com>"]
-version = "0.0.1"
-license = "Apache-2.0"
-repository = "https://github.com/cloudflare/workers-rs/tree/main/libworker"
-description = "Idiomatic framework / abstractions over `edgeworker-sys` FFI bindings."
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-async-trait = "0.1.50"
-chrono = { version = "0.4", default-features = false }
-chrono-tz = { version = "0.4", default-features = false }
-edgeworker-sys = { path = "../edgeworker-sys", version = "0.0.1" }
-futures = "0.3.16"
-http = "0.2.4"
-js-sys = "0.3.53"
-matchit = "0.4.2"
-serde = { version = "1.0.126", features = ["derive"] }
-serde_json = "1.0.64"
-wasm-bindgen = "0.2.76"
-wasm-bindgen-futures = "0.4.24"
-worker-kv = "0.3.0"
-url = "2.2.2"
