+++ worker-build/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
 authors = ["Cloudflare Workers Team <workers@cloudflare.com>"]
 edition = "2018"
-name    = "rust-worker-build"
-version = "0.0.2"
+name    = "worker-build"
+version = "0.0.3"
 license = "Apache-2.0"
-repository = "https://github.com/cloudflare/workers-rs/tree/main/rust-worker-build"
+repository = "https://github.com/cloudflare/workers-rs/tree/main/worker-build"
 readme = "README.md"
 description = "This is a tool to be used as a custom build command for a Cloudflare Workers `workers-rs` project."
