+++ worker-macros/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "workers-rs-macros"
-version = "0.0.1"
+name = "worker-macros"
+version = "0.0.2"
 authors = ["Cloudflare Workers Team <workers@cloudflare.com>"]
-repository = "https://github.com/cloudflare/workers-rs/tree/main/macros"
+repository = "https://github.com/cloudflare/workers-rs/tree/main/worker-macros"
 edition = "2018"
 license = "Apache-2.0"
-description = "Marcos to support the `worker` crate and remove FFI boilerplate / conversion from user code."
+description = "Macros to support the `worker` crate and remove FFI boilerplate / conversion from user code."
 
 [lib]
 proc-macro = true
@@ -13,8 +13,7 @@ path = "src/lib.rs"
 
 [dependencies]
 async-trait = "0.1.51"
-edgeworker-sys = { path = "../edgeworker-sys", version = "0.0.1" }
-libworker = { path = "../libworker", version = "0.0.1" }
+worker-sys = { path = "../worker-sys", version = "0.0.2" }
 syn = "1.0.74"
 proc-macro2 = "1.0.28"
 quote = "1.0.9"
