+++ worker-macros/src/event.rs
@@ -27,7 +27,7 @@ pub fn expand_macro(attr: TokenStream, item: TokenStream) -> TokenStream {
         }
     }
     let handler_type = handler_type
-        .expect("must have either 'fetch' or 'scheduled' attribute, e.g. #[cf::worker(fetch)]");
+        .expect("must have either 'fetch' or 'scheduled' attribute, e.g. #[event(fetch)]");
 
     // create new var using syn item of the attributed fn
     let mut input_fn = parse_macro_input!(item as ItemFn);
@@ -55,12 +55,12 @@ pub fn expand_macro(attr: TokenStream, item: TokenStream) -> TokenStream {
                 }
             };
 
-            // create a new "main" function that takes the edgeworker_sys::Request, and calls the
+            // create a new "main" function that takes the worker_sys::Request, and calls the
             // original attributed function, passing in a converted worker::Request
             let wrapper_fn = quote! {
-                pub async fn #wrapper_fn_ident(req: edgeworker_sys::Request, env: ::worker::Env) -> edgeworker_sys::Response {
+                pub async fn #wrapper_fn_ident(req: worker_sys::Request, env: ::worker::Env) -> worker_sys::Response {
                     // get the worker::Result<worker::Response> by calling the original fn
-                    match #input_fn_ident(worker::Request::from(req), env).await.map(edgeworker_sys::Response::from) {
+                    match #input_fn_ident(worker::Request::from(req), env).await.map(worker_sys::Response::from) {
                         Ok(res) => res,
                         Err(e) => {
                             ::worker::console_log!("{}", &e);
