+++ worker-sandbox/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["Cloudflare Workers Team <workers@cloudflare.com>"]
 edition = "2018"
-name = "rust-sandbox"
+name = "worker-sandbox"
 version = "0.1.0"
 license = "Apache-2.0"
 
@@ -23,7 +23,6 @@ http = "0.2.4"
 regex = "1.5.4"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-url = "2.2.2"
 wee_alloc = { version = "0.4.2", optional = true }
 worker = { path = "../worker" }
