+++ worker-sys/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "edgeworker-sys"
+name = "worker-sys"
 authors = ["Cloudflare Workers Team <workers@cloudflare.com>"]
 edition = "2018"
-version = "0.0.1"
+version = "0.0.2"
 license = "Apache-2.0"
-repository = "https://github.com/cloudflare/workers-rs/tree/main/edgeworker-sys"
+repository = "https://github.com/cloudflare/workers-rs/tree/main/worker-sys"
 description = "Low-level extern definitions / FFI bindings to the Cloudflare Workers JS Runtime."
 
 [dependencies]
