+++ worker/src/global.rs
@@ -1,5 +1,5 @@
 use crate::{request::Request as WorkerRequest, response::Response as WorkerResponse, Result};
-use edgeworker_sys::{Response as EdgeResponse, WorkerGlobalScope};
+use worker_sys::{Response as EdgeResponse, WorkerGlobalScope};
 
 use wasm_bindgen::JsCast;
 use wasm_bindgen_futures::JsFuture;
