+++ worker/src/lib.rs
@@ -1,11 +1,41 @@
-pub use libworker::durable;
-pub use libworker::durable::*;
-pub use libworker::prelude::*;
-pub use workers_rs_macros::{durable_object, event};
+mod cf;
+mod date;
+pub mod durable;
+mod env;
+mod error;
+mod formdata;
+mod global;
+mod headers;
+mod http;
+mod request;
+mod request_init;
+mod response;
+mod router;
+
+use std::result::Result as StdResult;
+
+pub type Result<T> = StdResult<T, error::Error>;
+
+pub use crate::date::{Date, DateInit};
+pub use crate::env::Env;
+pub use crate::error::Error;
+pub use crate::formdata::*;
+pub use crate::global::Fetch;
+pub use crate::headers::Headers;
+pub use crate::http::Method;
+pub use crate::request::Request;
+pub use crate::request_init::*;
+pub use crate::response::Response;
+pub use crate::router::{RouteContext, RouteParams, Router};
+pub use cf::Cf;
+
+pub use worker_sys::console_log;
+
+pub use crate::durable::*;
+pub use worker_macros::{durable_object, event};
 
 pub use async_trait;
-pub use edgeworker_sys;
 pub use js_sys;
-pub use libworker;
 pub use wasm_bindgen;
 pub use wasm_bindgen_futures;
+pub use worker_sys;
